update: routes.js new conditions
Checking if task, body and description exist

+++ 01-Challenge/src/routes.js
@@ -19,6 +19,9 @@ export const routes = [
 		handler: (req, res) => {
 			const { title, description } = req.body
 
+			if (!title || !description) 
+				return res.writeHead(400).end(JSON.stringify({ error: 'Title and description are required!' }))
+
 			const task = {
 				id: crypto.randomUUID(),
 				title,
@@ -38,13 +41,20 @@ export const routes = [
 		handler: (req, res) => {
 			const { id } = req.params
 			const { title, description } = req.body
+			const [task] = database.select('tasks', { id })
+
+			if (!task)
+				return res.writeHead(404).end(JSON.stringify({ error: 'Task not found!' }))
+
+			if (!title || !description) 
+				return res.writeHead(400).end(JSON.stringify({ error: 'Title and description are required!' }))			
 
 			database.update('tasks', id, {
 				title,
 				description,
 				updated_at: new Date(),
 			})
-			return res.writeHead(204).end('Update task!')
+			return res.writeHead(204).end()
 		}
 	},
 	{
@@ -52,8 +62,13 @@ export const routes = [
 		path: buildRoutePath('/tasks/:id'),
 		handler: (req, res) => {
 			const { id } = req.params
+			const [task] = database.select('tasks', { id })
+
+			if (!task)
+				return res.writeHead(404).end(JSON.stringify({ error: 'Task not found!' }))
+
 			database.delete('tasks', id)
-			return res.writeHead(204).end('Delete task!')
+			return res.writeHead(204).end()
 		}
 	},
 	{
@@ -61,10 +76,19 @@ export const routes = [
 		path: buildRoutePath('/tasks/:id/complete'),
 		handler: (req, res) => {
 			const { id } = req.params
+
+			const [task] = database.select('tasks', { id })
+
+			if (!task)
+				return res.writeHead(404).end(JSON.stringify({ error: 'Task not found!' }))
+
+			if (task.completed_at !== null)
+				return res.writeHead(400).end(JSON.stringify({ error: 'Task already completed!' }))
+
 			database.update('tasks', id, {
 				completed_at: new Date(),
 			})
-			return res.writeHead(204).end('Complete task!')
+			return res.writeHead(204).end()
 		}
 	}
 ]
